add new exercise

+++ compro1/FOR LOOP/triangle/triangle6.c
@@ -2,7 +2,7 @@
 
 int main(int argc, char const *argv[]) {
   for (int i = 5; i >= 1; i--) {
-    for (size_t j = 1; j <= i; j++) {
+    for (int j = 1; j <= i; j++) {
       printf("%d",j );
     }
     printf("\n");
